feat: user access features added in dashboard

+++ backend/routes/employees.js
@@ -3,8 +3,8 @@ const router = express.Router();
 const db = require('../db');
 const { verifyToken, isAdmin } = require('../middleware/auth');
 
-// Get all employees (ADMIN only)
-router.get('/', verifyToken, isAdmin, async(req, res) => {
+// Get all employees
+router.get('/', verifyToken, async(req, res) => {
     try{
         const results = await db.query('SELECT * FROM employees ORDER BY id ')
         res.json(results.rows)
@@ -16,10 +16,13 @@ router.get('/', verifyToken, isAdmin, async(req, res) => {
 
 
 //Add employee (ADMIN only)
-router.post('/', verifyToken, isAdmin, async(req, res) => {
+router.post('/', verifyToken, async(req, res) => {
 
     const {name, email, position, department, salary} = req.body
     try{
+        if (req.user.role !== 'ADMIN'){
+            return res.status(403).json({ error: 'Access denied, admin only'})
+        }
         await db.query( `INSERT INTO employees(name, email, position, department, salary) VALUES ($1, $2, $3, $4, $5)`,
         [name, email, position, department, salary])
         res.json({message: 'Employee added successfully'})
@@ -30,8 +33,8 @@ router.post('/', verifyToken, isAdmin, async(req, res) => {
     }
 });
 
-// Get a single employee by ID (ADMIN only)
-router.get('/:id', verifyToken, isAdmin, async (req, res) => {
+// Get a single employee by ID 
+router.get('/:id', verifyToken, async (req, res) => {
   const { id } = req.params;
   try {
     const result = await db.query('SELECT * FROM employees WHERE id = $1', [id]);
@@ -47,10 +50,13 @@ router.get('/:id', verifyToken, isAdmin, async (req, res) => {
 
 
 // Update employee (ADMIN only)
-router.put('/:id', verifyToken, isAdmin, async(req, res) =>{
+router.put('/:id', verifyToken, async(req, res) =>{
     const {id} = req.params
     const {name, email, position, department, salary} = req.body
     try{
+        if (req.user.role !== 'ADMIN'){
+            return res.status(403).json({ error: 'Access denied, admin only'})
+        }
         await db.query(` UPDATE employees SET name=$1, email=$2, position=$3, department=$4, salary=$5 WHERE id=$6`,
             [name, email, position, department, salary, id])
         res.json({message: 'Employee updated successfully'})
@@ -61,9 +67,12 @@ router.put('/:id', verifyToken, isAdmin, async(req, res) =>{
 })
 
 // Delete employee (ADMIN only)
-router.delete('/:id', verifyToken, isAdmin, async(req, res) =>{
+router.delete('/:id', verifyToken, async(req, res) =>{
     const {id} = req.params     
     try{
+        if (req.user.role !== 'ADMIN'){
+            return res.status(403).json({ error: 'Access denied, admin only'})
+        }
         await db.query('DELETE FROM employees WHERE id=$1', [id])
         res.json({message: 'Employee deleted successfully'})
     }catch(err){
